requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.9
commonmark==0.9.1
config==0.5.1
numpy==1.23.3
Pygments==2.13.0
rich==12.5.1
RPi.GPIO==0.7.1
smbus==1.1.post2
spidev==3.5
wiringpi==2.60.1
